+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+- New option `expand_tabs` in `Style` causes tab titles to expand to fill the width of their tab bars.
+- Added `context_menu` into `TabViewer` ([#46](https://github.com/Adanos020/egui_dock/pull/46)).        
+- The `ScrollArea` inside a Tab is now Optional via `Style` ([#49](https://github.com/Adanos020/egui_dock/pull/49)).
+- `Tree::tabs`: an iterator over the tabs in a tree ([#53](https://github.com/Adanos020/egui_dock/pull/53)).
 
 ## 0.2.1 - 2022-09-09